Prepare system to solve numpy overflow problem

+++ Bridge.py
@@ -1,4 +1,5 @@
 import random
+import numpy as np
 
 import Neuron
 
@@ -10,8 +11,8 @@ class Bridge:
     self.__StartNeuron = StartNeuron
     self.__FinishNeuron = FinishNeuron
 
-    if(len(args) > 0): self.Weight = args[0]                           #Se il valore del peso è già scelto
-    else: self.Weight = random.randint(-10000, 10000) / 1000           #Valore scelto per convenienza da -1 ad 1 ma si può cambiare il range volendo
+    if(len(args) > 0): self.Weight = args[0]                                   #Se il valore del peso è già scelto
+    else: self.Weight = np.float64(random.randint(-10000, 10000) / 1000)       #Valore scelto per convenienza da -1 ad 1 ma si può cambiare il range volendo
 
   def GetStartNeuron(self):
     return self.__StartNeuron
